% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{calculate_permutation_testing}
\alias{calculate_permutation_testing}
\title{calculate_permutation_testing}
\usage{
calculate_permutation_testing(
  baseline_pops,
  model_pops,
  number_of_permutations,
  clinvar
)
}
\arguments{
\item{baseline_pops}{baseline population}

\item{model_pops}{model population}

\item{number_of_permutations}{number of permitations}
}
\description{
Permutation testing for significance of predictor
}
