% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{calculatePermutationTesting}
\alias{calculatePermutationTesting}
\title{calculatePermutationTesting}
\usage{
calculatePermutationTesting(
  baseline.pops,
  model.pops,
  numberOfPermutations,
  clinvar
)
}
\description{
Permutation testing for significance of predictor
}
