% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_hard_error}
\alias{get_hard_error}
\title{get_hard_error}
\usage{
get_hard_error(model, clinvar)
}
\arguments{
\item{model}{model}

\item{clinvar}{clinvar data}
}
\description{
Hard missclassification error
}
