% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_soft_error}
\alias{get_soft_error}
\title{get_soft_error}
\usage{
get_soft_error(model, clinvar)
}
\arguments{
\item{model}{model}

\item{clinvar}{clinvar data}
}
\description{
Soft missclassification error
}
