% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_tobias}
\alias{launch_tobias}
\title{Launch Tobias}
\usage{
launch_tobias(browser = TRUE, ...)
}
\arguments{
\item{...}{Additional parameters passed to \code{\link[shiny]{runApp}}}

\item{launch.browser}{Launch in browser? Default TRUE}
}
\description{
Launch Tobias
}
