% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{plot_ancestry_effects}
\alias{plot_ancestry_effects}
\title{plot_ancestry_effects}
\usage{
plot_ancestry_effects(mypops, popLabel, popColor, model)
}
\description{
Contains logic to PLOT the regression models. Plot for Global + Ancestry Effects
}
