% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{plot_global_effects}
\alias{plot_global_effects}
\title{plot_global_effects}
\usage{
plot_global_effects(test_global_f)
}
\description{
Contains logic to PLOT the regression models. Plot for Global Effects
}
