% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{plotAncestryEffects}
\alias{plotAncestryEffects}
\title{plotAncestryEffects}
\usage{
plotAncestryEffects(mypops, popLabel, popColor, model)
}
\description{
Contains logic to PLOT the regression models. Plot for Global + Ancestry Effects
}
