% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{plotPopEffects}
\alias{plotPopEffects}
\title{plotPopEffects}
\usage{
plotPopEffects(mypops, popLabel, popColor, model)
}
\description{
Contains logic to PLOT the regression models. Plot for Global + Population Effects
}
