% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{readData}
\alias{readData}
\title{/Users/snehit/dev/ancestrybias/clinvar.exac.variants.gene.submission.diseases.alleles.tab'}
\usage{
readData(path)
}
\description{
/Users/snehit/dev/ancestrybias/clinvar.exac.variants.gene.submission.diseases.alleles.tab'
}
