% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurize_AFs.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter_plot}
\usage{
scatter_plot(
  dataset,
  feature1,
  feature2,
  colorByClass,
  colorcode,
  title,
  xlabel,
  ylabel
)
}
\description{
scatter_plot
}
