% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{calculate_permutation_testing}
\alias{calculate_permutation_testing}
\title{calculate_permutation_testing}
\usage{
calculate_permutation_testing(
  baseline.pops,
  model.pops,
  numberOfPermutations,
  clinvar
)
}
\description{
Permutation testing for significance of predictor
}
