% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{get_z_score}
\alias{get_z_score}
\title{get_z_score}
\usage{
get_z_score(model)
}
\arguments{
\item{model}{model used for Z scoring}
}
\description{
Contains logic to TEST the regression models (permutation tests, statistical significance, prediction power, etc.)
}
