% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{plot_pop_effects}
\alias{plot_pop_effects}
\title{plot_pop_effects}
\usage{
plot_pop_effects(mypops, popLabel, popColor, model)
}
\arguments{
\item{mypops}{populations}

\item{popLabel}{populations labels}

\item{popColor}{populations colors}

\item{model}{model used for effects estimation}
}
\description{
Contains logic to PLOT the regression models. Plot for Global + Population Effects
}
